requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
atoma==0.0.17
attrs==21.4.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.2.1
favicon==0.7.0
feedparser==6.0.10
Flask==2.1.2
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
Flask-SQLAlchemy==2.5.1
Flask-thumbnails==1.1.0
Flask-WTF==1.0.1
flaskfilemanager==0.0.6
frozenlist==1.3.0
geoip2==4.6.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
IPy==1.1
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
littlefish==0.0.70
lxml==4.9.0
MarkupSafe==2.1.1
maxminddb==2.2.0
mongoengine==0.24.1
multidict==6.0.2
passlib==1.7.4
phonenumbers==8.12.51
Pillow==9.1.1
pycparser==2.21
PyMarkovChain==1.8
pymongo==4.1.1
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
speaklater==1.3
SQLAlchemy==1.4.39
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.2
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
